Remove unused delete assignment endpoint, added create entity and update space and refactoring
