+++ api/src/main/kotlin/com/ford/internalprojects/peoplemover/baserepository/PeopleMoverRepositoryImpl.kt
@@ -17,6 +17,7 @@
 
 package com.ford.internalprojects.peoplemover.baserepository
 
+import com.ford.internalprojects.peoplemover.baserepository.exceptions.EntityAlreadyExistsException
 import com.ford.internalprojects.peoplemover.baserepository.exceptions.EntityNotExistsException
 import com.ford.internalprojects.peoplemover.space.SpaceComponent
 import com.ford.internalprojects.peoplemover.space.SpaceRepository
@@ -24,8 +25,6 @@ import com.ford.internalprojects.peoplemover.space.exceptions.SpaceNotExistsExce
 import org.springframework.data.jpa.repository.support.JpaEntityInformation
 import org.springframework.data.jpa.repository.support.SimpleJpaRepository
 import org.springframework.data.repository.findByIdOrNull
-import java.io.Serializable
-import java.lang.RuntimeException
 import java.sql.Timestamp
 import java.util.*
 import javax.persistence.EntityManager
@@ -44,16 +43,31 @@ class PeopleMoverRepositoryImpl<T : SpaceComponent, ID : Int>(
     }
 
     @Transactional
+    override fun <S : T> createEntityAndUpdateSpaceLastModified(entity: S): S {
+        if (entity.id != null) {
+            throw EntityAlreadyExistsException()
+        } else {
+            updateSpaceLastModified(entity.spaceUuid)
+            return save(entity)
+        }
+    }
+
+    @Transactional
     override fun <S : T> updateEntityAndUpdateSpaceLastModified(entity: S): S {
         val entityToUpdate =  findByIdOrNull(entity.id!!)
         if(entityToUpdate == null || entityToUpdate.spaceUuid != entity.spaceUuid) {
             throw EntityNotExistsException()
         }
-        return saveAndUpdateSpaceLastModified(entity)
+        updateSpaceLastModified(entity.spaceUuid)
+        return save(entity)
     }
 
     @Transactional
-    override fun <S : T> deleteAndUpdateSpaceLastModified(entity: S) {
+    override fun <S : T> deleteEntityAndUpdateSpaceLastModified(entity: S) {
+        val entityToDelete =  findByIdOrNull(entity.id!!)
+        if(entityToDelete == null || entityToDelete.spaceUuid != entity.spaceUuid) {
+            throw EntityNotExistsException()
+        }
         updateSpaceLastModified(entity.spaceUuid)
         delete(entity)
     }
